requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
google-generativeai==0.3.2
Pillow==9.5.0
PyMuPDF==1.23.8
python-docx==1.1.0
weasyprint==60.2
gunicorn==21.2.0
markdown==3.5.1
